requirements.txt
ffmpeg
google.cloud.texttospeech
gtts
py-cord
pynacl




